scripts/requirements.txt
requests>=2.31.0
python-frontmatter>=1.1.0
tweepy>=4.14.0
